fix: standardize playlist results
