+++ src/controllers/user/playlists/routes/{id}/index.js
@@ -1,7 +1,7 @@
 const { User, Playlist, PlaylistItem, Track, File } = require('../../../../../db/models')
 const { Op } = require('sequelize')
-const coverSrc = require('../../../../../util/cover-src')
 const { authenticate } = require('../../../authenticate')
+const playlistService = require('../../../../playlists/services/playlistService')
 
 module.exports = function () {
   const operations = {
@@ -88,10 +88,10 @@ module.exports = function () {
     try {
       let result = await Playlist.findOne({
         attributes: [
-          'artistId'
+          'id'
         ],
         where: {
-          userId: ctx.profile.id,
+          creatorId: ctx.profile.id,
           id: ctx.params.id
         }
       })
@@ -103,7 +103,7 @@ module.exports = function () {
 
       result = await Playlist.update(body, {
         where: {
-          userId: ctx.profile.id,
+          creatorId: ctx.profile.id,
           id: ctx.params.id
         },
         returning: true
@@ -125,13 +125,13 @@ module.exports = function () {
           'title'
         ],
         where: {
-          userId: ctx.profile.id,
+          creatorId: ctx.profile.id,
           id: ctx.params.id
         }
       })
 
       ctx.body = {
-        data: result,
+        data: playlistService(ctx).single(result),
         status: 'ok'
       }
     } catch (err) {
@@ -192,6 +192,7 @@ module.exports = function () {
           'about',
           'id',
           'private',
+          'creatorId',
           'tags',
           'cover',
           'title'
@@ -264,76 +265,8 @@ module.exports = function () {
         plain: true
       })
 
-      let ext = '.jpg'
-
-      if (ctx.accepts('image/webp')) {
-        ext = '.webp'
-      }
-
-      const variants = [120, 600, 1500]
-
       ctx.body = {
-        data: {
-          about: data.about,
-          cover: coverSrc(data.cover, !data.cover_metadata ? '600' : '1500', ext, !data.cover_metadata),
-          cover_metadata: {
-            id: data.cover
-          },
-          artistId: data.artistId,
-          display_artist: data.display_artist,
-          user: {
-            id: data.user.id
-          },
-          download: data.download,
-          id: data.id,
-          items: data.items.map((item) => {
-            const fallback = !item.track.cover_art ? false : !item.track.cover_metadata
-
-            return {
-              index: item.index,
-              track: {
-                id: item.track.id,
-                title: item.track.title,
-                status: item.track.status,
-                album: item.track.album,
-                duration: item.track.duration,
-                creator_id: item.track.creator_id,
-                artist: item.track.artist,
-                cover: coverSrc(item.track.cover_art || data.cover, '600', ext, fallback),
-                images: variants.reduce((o, key) => {
-                  const variant = ['small', 'medium', 'large'][variants.indexOf(key)]
-
-                  return Object.assign(o,
-                    {
-                      [variant]: {
-                        width: key,
-                        height: key,
-                        url: coverSrc(item.track.cover_art || data.cover, key, ext, fallback)
-                      }
-                    }
-                  )
-                }, {}),
-                url: `${process.env.APP_HOST}/api/v3/user/stream/${item.track.id}`
-              }
-            }
-          }),
-          images: variants.reduce((o, key) => {
-            const variant = ['small', 'medium', 'large'][variants.indexOf(key)]
-
-            return Object.assign(o,
-              {
-                [variant]: {
-                  width: key,
-                  height: key,
-                  url: coverSrc(data.cover, key, ext, !data.cover_metadata)
-                }
-              }
-            )
-          }, {}),
-          private: data.private,
-          tags: data.tags,
-          title: data.title
-        },
+        data: playlistService(ctx).single(data),
         status: 'ok'
       }
     } catch (err) {
